IMS: Add handling for multianchor conference.
Just based on multiparty bit we should not determine
merge peer or merge host. We should check for conference host
to determine merge peer or merge host. This will ensure we
determine correct merge peer and merge host even when merge
is clicked on participant side where multiparty is true but it is
not a conference host.

Change-Id: I463accb767f8d74801af01646b962caee452c231
CRs-Fixed: 2556697